pyproject.toml
[tool.poetry]
name = "ehp-ml-challenge"
version = "0.1.0"
description = ""
authors = ["HyeonJin <choihj94@huray.net>"]
readme = "README.md"
packages = [{include = "ehp_ml_challenge"}]

[tool.poetry.dependencies]
python = "^3.10"
dependency-injector = "^4.41.0"
torch = "^1.13.1"
torchvision = "^0.14.1"
albumentations = "^1.3.0"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
